errors fixed

+++ SenshiProject/src/main/java/com/example/SenshiProject/controller/CardController.java
@@ -13,32 +13,32 @@ public class CardController {
     CardService cardService;
 
     @PostMapping("/issueCard/{accountId}")
-    public Card issueCard(@PathVariable Long accountId){
+    public Card issueCard(@PathVariable int accountId){
         return cardService.issueCard(accountId);
     }
 
     @GetMapping("/{accountId/getCards}")
-    public List<Card> getCards(@PathVariable Long accountId){
+    public List<Card> getCards(@PathVariable int accountId){
         return cardService.getCardsByAccount(accountId);
     }
 
     @PutMapping("/{cardId}/activateCard")
-    public Card activateCard(@PathVariable Long cardId){
+    public Card activateCard(@PathVariable int  cardId){
         return cardService.activateCard(cardId);
     }
 
     @PutMapping("/{cardId}/blockCard")
-    public Card blockCard(@PathVariable Long cardId){
+    public Card blockCard(@PathVariable int cardId){
         return cardService.blockCard(cardId);
     }
 
     @PutMapping("/{cardId}/unblockCard")
-    public Card unblockCard(@PathVariable Long cardId){
+    public Card unblockCard(@PathVariable int  cardId){
         return cardService.unblockCard(cardId);
     }
 
     @PutMapping("/{cardId}/closeCard")
-    public String closeCard(Long cardId){
+    public String closeCard(int  cardId){
         return cardService.closeCard(cardId);
     }
 }
