+++ SenshiProject/src/main/java/com/example/SenshiProject/serviceinterface/CardServiceInterface.java
@@ -6,10 +6,10 @@ import com.example.SenshiProject.entity.Card;
 import java.util.List;
 
 public interface CardServiceInterface {
-     Card issueCard(Long accountId);
-     List<Card> getCardsByAccount(Long accountId);
-     Card activateCard(Long cardId);
-     Card blockCard(Long cardId);
-     Card unblockCard(Long cardId);
-     String closeCard(Long cardId);
+     Card issueCard(int  accountId);
+     List<Card> getCardsByAccount(int  accountId);
+     Card activateCard(int  cardId);
+     Card blockCard(int cardId);
+     Card unblockCard(int  cardId);
+     String closeCard(int  cardId);
 }
